+++ roslibrust/src/lib.rs
@@ -41,9 +41,19 @@ use tungstenite::Message;
 
 /// Used for type erasure of message type so that we can store arbitrary handles
 type Callback = Box<dyn Fn(&str) -> () + Send + Sync>;
+/// Type erasure of callback for a service
+/// Internally this will covert the input string to the Request type
+/// Send that converted type into the user's callback
+/// Get the result of the user's callback and then serialize that so it can be transmitted
+// TODO need to implement drop logic on this
+type ServiceCallback = Box<
+    dyn Fn(&str) -> Result<serde_json::Value, Box<dyn std::error::Error + Send + Sync>>
+        + Send
+        + Sync,
+>;
 
-/// Our underlying comunication type
-//TODO we really want to split this into seperate read/write heads that can be locked seperatly
+/// Our underlying communication type
+//TODO we really want to split this into separate read/write heads that can be locked separately
 type Socket = tokio_tungstenite::WebSocketStream<tokio_tungstenite::MaybeTlsStream<TcpStream>>;
 type Reader = SplitStream<Socket>;
 type Writer = SplitSink<Socket, Message>;
@@ -111,6 +121,17 @@ impl RosMessageType for () {
     const ROS_TYPE_NAME: &'static str = "";
 }
 
+/// Fundamental traits for service types this crate works with
+/// This trait will be satisfied for any services definitions generated with this crate's message_gen functionality
+pub trait RosServiceType {
+    /// Name of the ros service e.g. `rospy_tutorials/AddTwoInts`
+    const ROS_SERVICE_NAME: &'static str;
+    /// The type of data being sent in the request
+    type Request: RosMessageType;
+    /// The type of the data
+    type Response: RosMessageType;
+}
+
 /// Builder options for creating a client
 #[derive(Clone)]
 pub struct ClientOptions {
@@ -145,6 +166,7 @@ pub struct Client {
     // Stores a record of the publishers we've handed out
     publishers: Arc<DashMap<String, PublisherHandle>>,
     subscriptions: Arc<DashMap<String, Subscription>>,
+    services: Arc<DashMap<String, ServiceCallback>>,
     // Contains any outstanding service calls we're waiting for a response on
     // Map key will be a uniquely generated id for each call
     service_calls: Arc<DashMap<String, tokio::sync::oneshot::Sender<Value>>>,
@@ -159,6 +181,7 @@ impl Client {
             reader: Arc::new(RwLock::new(reader)),
             writer: Arc::new(RwLock::new(writer)),
             publishers: Arc::new(DashMap::new()),
+            services: Arc::new(DashMap::new()),
             subscriptions: Arc::new(DashMap::new()),
             service_calls: Arc::new(DashMap::new()),
             opts: Arc::new(opts),
@@ -338,16 +361,44 @@ impl Client {
         });
     }
 
-    // TODO apparently rosbridge doesn't support this!
-    // TODO get them to! and then implement!
-    // /// Advertises a service and returns a handle to the service server
-    // /// Serivce will be active until the handle is dropped!
-    // pub async fn advertise_service(
-    //     &mut self,
-    //     topic: &str,
-    // ) -> Result<ServiceServer, Box<dyn Error>> {
-    //     unimplemented!()
-    // }
+    /// Advertises a service and returns a handle to the service server
+    /// Service will be active until the handle is dropped!
+    pub async fn advertise_service<T: RosServiceType>(
+        &mut self,
+        topic: &str,
+        server: fn(
+            T::Request,
+        )
+            -> Result<T::Response, Box<dyn std::error::Error + 'static + Send + Sync>>,
+    ) -> RosLibRustResult<()> {
+        let mut writer = self.writer.write().await;
+        writer.advertise_service(topic, T::ROS_SERVICE_NAME).await?;
+
+        // We need to do type erasure and hide the request by wrapping their closure in a generic closure
+        let erased_closure = move |message: &str| -> Result<
+            serde_json::Value,
+            Box<dyn std::error::Error + Send + Sync>,
+        > {
+            // Type erase the incoming type
+            let parsed_msg = serde_json::from_str(message)?;
+            let response = server(parsed_msg)?;
+            // Type erase the outgoing type
+            let response_string = serde_json::json!(response);
+            Ok(response_string)
+        };
+
+        let res = self
+            .services
+            .insert(topic.to_string(), Box::new(erased_closure));
+        if let Some(_previous_server) = res {
+            warn!("Re-registering a server for a pre-existing topic? Are you sure you want to do this");
+            unimplemented!()
+        }
+
+        //TODO need to provide unadvertise and drop callback
+
+        Ok(())
+    }
 
     /// Calls a ros service and returns the response
     ///
@@ -370,15 +421,26 @@ impl Client {
             .collect();
         {
             if self.service_calls.insert(rand_string.clone(), tx).is_some() {
-                error!("ID collision encounted in call_service");
+                error!("ID collision encountered in call_service");
             }
         }
         {
             let mut comm = self.writer.write().await;
-            comm.call_service(service, &rand_string, req).await?;
+            Client::timeout(
+                self.opts.timeout,
+                comm.call_service(service, &rand_string, req),
+            )
+            .await?;
         }
-        // TODO timeout impl here
-        let msg = match rx.await {
+
+        // Having to do manual timeout logic here because of error types
+        let recv = if let Some(timeout) = self.opts.timeout {
+            tokio::time::timeout(timeout, rx).await?
+        } else {
+            rx.await
+        };
+
+        let msg = match recv {
             Ok(msg) => msg,
             Err(e) =>
             // TODO remove panic! here, this could result from dropping communication, need to handle disconnect better
@@ -437,10 +499,16 @@ impl Client {
                     let op = comm::Ops::from_str(op)?;
                     match op {
                         comm::Ops::Publish => {
+                            trace!("handling publish for {:?}", &parsed);
                             self.handle_publish(parsed).await;
                         }
                         comm::Ops::ServiceResponse => {
+                            trace!("handling service response for {:?}", &parsed);
                             self.handle_response(parsed).await;
+                        }
+                        comm::Ops::CallService => {
+                            trace!("handling call_service for {:?}", &parsed);
+                            self.handle_service(parsed).await;
                         }
                         _ => {
                             warn!("Unhandled op type {}", op)
@@ -511,15 +579,50 @@ impl Client {
             .get(data.get("topic").unwrap().as_str().unwrap());
         let callbacks = match callbacks {
             Some(callbacks) => callbacks,
-            _ => panic!("Received publish message for unsubscribed topic!"),
+            _ => panic!("Received publish message for unsubscribed topic!"), // TODO probably shouldn't be a panic?
         };
         for (_id, callback) in &callbacks.handles {
             callback(
+                // TODO possible bug here if "msg" isn't defined remove this unwrap
                 serde_json::to_string(data.get("msg").unwrap())
                     .unwrap()
                     .as_str(),
             )
         }
+    }
+
+    /// Response handler for receiving a service call looks up if we have a service
+    /// registered for the incoming topic and if so dispatches to the callback
+    async fn handle_service(&mut self, data: Value) {
+        // Unwrap is okay, field is fully required and strictly type
+        let topic = data.get("service").unwrap().as_str().unwrap();
+        // Unwrap is okay, field is strictly typed to string
+        let id = data.get("id").map(|id| id.as_str().unwrap().to_string());
+
+        // Lookup if we have a service for the message
+        let callback = self.services.get(topic);
+        let callback = match callback {
+            Some(callback) => callback,
+            _ => panic!("Received call_service for unadvertised service!"),
+        };
+        // TODO likely bugs here remove this unwrap. Unclear what we are expected to get for empty service
+        let request = data.get("args").unwrap().to_string();
+        let mut writer = self.writer.write().await;
+        match callback(&request) {
+            Ok(res) => {
+                // TODO unwrap here is probably bad... Failure to write means disconnected?
+                writer.service_response(topic, id, true, res).await.unwrap();
+            }
+            Err(e) => {
+                error!("A service callback on topic {:?} failed with {:?} sending response false in service_response", data.get("service"), e);
+                writer
+                    .service_response(topic, id, false, serde_json::json!(format!("{e}")))
+                    .await
+                    .unwrap();
+            }
+        };
+
+        // Now we need to send the service_response back
     }
 
     async fn handle_response(&mut self, data: Value) {
@@ -548,7 +651,7 @@ impl Client {
         }
     }
 
-    /// Bassic connection attempt and error wrapping
+    /// Basic connection attempt and error wrapping
     async fn connect(url: &str) -> RosLibRustResult<Socket> {
         let attempt = tokio_tungstenite::connect_async(url).await;
         match attempt {